test: remove usage of 'remote' module from webview spec (#20048)
