+++ spec/fixtures/pages/permissions/geolocation.html
@@ -1,5 +1,5 @@
 <script>
 navigator.geolocation.getCurrentPosition(() => {}, (err) => {
-  require('electron').ipcRenderer.sendToHost('message', err.message);
+  require('electron').ipcRenderer.send('message', err.message);
 });
 </script>
