+++ spec/fixtures/pages/permissions/midi.html
@@ -1,5 +1,5 @@
 <script>
 navigator.requestMIDIAccess({sysex: false}).then(() => {}, (err) => {
-  require('electron').ipcRenderer.sendToHost('message', err.name);
+  require('electron').ipcRenderer.send('message', err.name);
 });
 </script>
