+++ spec/fixtures/pages/webview-in-page-navigate.html
@@ -1,6 +1,6 @@
 <html>
 <body>
-<webview nodeintegration src="zoom://host1" id="view" partition="webview-temp"/>
+<webview nodeintegration src="app://host1" id="view" partition="webview-temp"/>
 </body>
 <script>
   const {ipcRenderer, webFrame} = require('electron')
